requirements.txt
# Web Framework and Database
Flask>=2.0.1
Flask-SQLAlchemy>=3.0.2
Flask-Migrate>=4.0.4
Flask-Login>=0.6.2
SQLAlchemy>=1.4.23

# Web Scraping and Automation
selenium>=4.1.0
beautifulsoup4>=4.9.3
webdriver-manager>=3.8.5
requests>=2.26.0
aiohttp>=3.8.1

# Data Processing
pandas>=1.3.0
numpy>=1.21.0
python-dateutil>=2.8.2
openpyxl>=3.0.9

# Progress and CLI
tqdm>=4.65.0
click>=8.0.3
colorama>=0.4.4

# Security and Environment
python-dotenv>=0.19.0
cryptography>=3.4.7
PyJWT>=2.3.0

# Error Handling and Logging
sentry-sdk>=1.5.1
python-json-logger>=2.0.2

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.1

# Documentation
Sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0

# Production
gunicorn>=20.1.0
supervisor>=4.2.4